refactor(cuda): split tactic identities from store
