+++ crates/apxinf-cuda/src/tuning/store.rs
@@ -3,87 +3,7 @@ use std::collections::HashMap;
 use apxinf_core::{Error, Result};
 
 use super::key::{GemmBucketKey, GemmTuningKey};
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum TacticBackend {
-    Cutlass,
-    CublasLt,
-    /// Fully specified cuBLASLt algorithm configuration. Unlike a heuristic
-    /// rank, this remains stable when the library reorders its candidates.
-    CublasLtCustom,
-    /// Fully specified cuBLASLt algorithm validated for the fused FP16-bias
-    /// operation descriptor used by static FP8 down/residual projections.
-    CublasLtCustomBias,
-    /// Fully specified cuBLASLt algorithm for two serial half-width GEMMs over
-    /// one packed weight/output allocation. This is an exact-shape backend:
-    /// it preserves the fused tensor layout while selecting a faster N/2
-    /// kernel family on shapes where the full-width kernel underperforms.
-    CublasLtCustomSplitSerial,
-    /// First-half cuBLASLt gate GEMM followed by a CUTLASS second-half up
-    /// GEMM whose epilogue performs exact GeGLU and E4M3 quantization. This
-    /// backend is valid only for explicitly tuned native-FP8 shapes.
-    CublasLtCustomSplitGeGluCutlass,
-    /// Two-SM 128x256x128 CUTLASS up/GeGLU epilogue using a 2x2
-    /// cluster. KernelScheduleAuto selects the SM100 two-SM, eight-stage path.
-    CublasLtCustomSplitGeGluCutlass2SmAuto,
-    /// Same two-SM cluster with an explicit three-stage
-    /// mainloop candidate. Kept separate so end-to-end timing can decide.
-    CublasLtCustomSplitGeGluCutlass2SmStage3,
-    /// Exact M522 language MLP path: current half-width
-    /// cuBLASLt gate tactic followed by an explicit SM100 two-SM CUTLASS up
-    /// GEMM with exact GeGLU and E4M3 quantization in its EVT epilogue.
-    CublasLtCustomSplitGeGluCutlassM522Explicit2Sm,
-    /// Exact-shape FP8 one-node dual-GEMM + GeGLU operator. Exact tuning keys
-    /// and runtime validation restrict it to the production M522/M533 shapes;
-    /// the resident weight must use the validated [gate256,up256] layout.
-    CutlassFp8DualGeGlu,
-    /// Exact BF16 M522 one-node dual-GEMM mega-kernel. The
-    /// resident weight must use the load-time-validated [gate256,up256]
-    /// physical layout.
-    CutlassBf16DualGeGluM522,
-    /// Exact BF16 M533 generalization of the accepted M522 one-node
-    /// dual-GEMM mega-kernel. It reuses the same
-    /// load-time-validated [gate256,up256] physical layout while remaining a
-    /// distinct exact-shape backend rather than a generic-M route.
-    CutlassBf16DualGeGluM533,
-    /// BF16 split-EVT first-half cuBLASLt gate followed by a native
-    /// SM100 CUTLASS up GEMM whose EVT preserves the exact BF16 GeGLU rounds.
-    CublasLtCustomSplitGeGluCutlassBf16,
-    Vendor,
-}
-
-/// Decoded representation of a compact `cublaslt_custom` tactic id.
-///
-/// Algorithm id 66, split-K=1, reduction=none, swizzle=0, and inner-shape=0
-/// are part of the backend contract. The remaining CUDA 13 configuration is
-/// packed into the signed JSON-compatible tactic value as follows:
-/// tile[9:0], custom[12:10], cluster[18:13], stages[24:19].
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct CublasLtCustomConfig {
-    pub tile_id: i32,
-    pub custom_option: i32,
-    pub cluster_shape_id: i32,
-    pub stages_id: i32,
-}
-
-pub fn decode_cublaslt_custom_tactic(value: i32) -> Option<CublasLtCustomConfig> {
-    if value <= 0 || value & !0x01ff_ffff != 0 {
-        return None;
-    }
-    let config = CublasLtCustomConfig {
-        tile_id: value & 0x3ff,
-        custom_option: (value >> 10) & 0x7,
-        cluster_shape_id: (value >> 13) & 0x3f,
-        stages_id: (value >> 19) & 0x3f,
-    };
-    (config.tile_id > 0 && config.stages_id > 0).then_some(config)
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub struct TacticId {
-    pub backend: TacticBackend,
-    pub value: i32,
-}
+use super::tactic::{TacticBackend, TacticId};
 
 #[derive(Clone, Debug, PartialEq)]
 pub struct GemmTuningRecord {
